clone_all_repos/requirements.txt
PyGithub>=1.55